Wifi: Use scan detailed cache when setting SSID allowlist for linked networks
ScanRequestProxy will not have the scan results corresponding
to the link AP if it is missed on the last scan or last scan
is not a full scan. This will cause the issue that the check
for FT-PSK/FT-SAE is skipped and SSID allowlist is set to
firmware for linked networks when the current connection is
FT-PSK/FT-SAE.

Use WifiConfigManager#getScanDetailCacheForNetwork(networkId)
instead to get scan detailed cache.

Change-Id: I57f7ed7712d7d19e6de4ebd1f6c6243c36486762
CRs-Fixed: 3391082